Move into MachineUIFactory class, add machine to builder
