+++ src/main/java/com/gregtechceu/gtceu/api/mui/factory/IMuiFactory.java
@@ -1,10 +1,11 @@
 package com.gregtechceu.gtceu.api.mui.factory;
 
+import com.gregtechceu.gtceu.api.machine.MetaMachine;
 import com.gregtechceu.gtceu.api.mui.base.IUIHolder;
 import com.gregtechceu.gtceu.api.mui.value.sync.PanelSyncManager;
 import com.gregtechceu.gtceu.client.mui.screen.ModularPanel;
 import com.gregtechceu.gtceu.client.mui.screen.UISettings;
-import com.gregtechceu.gtceu.common.mui.factory.BuilderMachineUIFactory;
+import com.gregtechceu.gtceu.common.mui.factory.MachineUIFactory;
 
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
@@ -16,7 +17,13 @@ import net.minecraft.world.phys.BlockHitResult;
 public interface IMuiFactory extends IUIHolder<PosGuiData> {
 
     @Override
-    ModularPanel buildUI(PosGuiData data, PanelSyncManager syncManager, UISettings settings);
+    default ModularPanel buildUI(PosGuiData data, PanelSyncManager syncManager, UISettings settings) {
+        var machine = MachineUIFactory.getMachine(data);
+        return buildUIFunction(data, syncManager, settings, machine);
+    };
+
+    ModularPanel buildUIFunction(PosGuiData data, PanelSyncManager syncManager, UISettings settings,
+                                 MetaMachine machine);
 
     default boolean shouldOpenUI(Player player, InteractionHand hand, BlockHitResult hit) {
         return true;
@@ -25,7 +32,7 @@ public interface IMuiFactory extends IUIHolder<PosGuiData> {
     default InteractionResult tryToOpenUI(Player player, InteractionHand hand, BlockHitResult hit) {
         if (this.shouldOpenUI(player, hand, hit)) {
             if (player instanceof ServerPlayer serverPlayer) {
-                BuilderMachineUIFactory.INSTANCE.open(serverPlayer, hit.getBlockPos(), this);
+                MachineUIFactory.INSTANCE.open(serverPlayer, hit.getBlockPos(), this);
             }
             return InteractionResult.sidedSuccess(player.level().isClientSide);
         } else {
